// types/types.ts
export interface SearchResult {
  id: string;
  type: 'player' | 'school' | 'parent' | 'guardian' | 'coach';
  name: string;
  gender?: string;
  grade?: string;
  dob?: string;
  aauNumber?: string;
  email?: string;
  phone?: string;
  address?: string | Address;
  image?: string;
  additionalInfo?: string;
  createdAt: string;
}

export interface Address {
  street: string;
  street2: string;
  city: string;
  state: string;
  zip: string;
}

export interface Parent {
  _id: string;
  password?: string;
  fullName: string;
  email: string;
  phone: string;
  address:
    | {
        street: string;
        street2: string;
        city: string;
        state: string;
        zip: string;
      }
    | string;
  relationship?: string;
  aauNumber?: string;
  role: string;
  isCoach?: boolean;
  players?: Player[];
  additionalGuardians?: Guardian[];
  createdAt?: string;
  updatedAt?: string;
  isGuardian?: boolean;
  avatar?: string;
  paymentComplete?: boolean;
}
// ...
export interface Guardian {
  id?: string;
  _id?: string;
  fullName: string;
  relationship: string;
  phone: string;
  email: string;
  address:
    | {
        street: string;
        street2: string;
        city: string;
        state: string;
        zip: string;
      }
    | string;
  isCoach: boolean;
  aauNumber: string;
  players?: Player[];
  createdAt?: string;
  updatedAt?: string;
}

export interface Player {
  _id: string;
  parentId?: string;
  fullName: string;
  gender: string;
  grade: string;
  dob: string;
  schoolName: string;
  healthConcerns: string;
  aauNumber: string;
  registrationYear: number;
  season: string;
  createdAt: string;
  updatedAt: string;
  status: string;
  __v: number;
  paymentComplete?: boolean;
}
// ...
export interface DecodedToken {
  id: string;
  role: 'admin' | 'coach' | 'user' | 'parent';
  exp?: number;
  email?: string;
  fullName?: string;
  phone?: string;
  address?: string | Address;
  relationship?: string;
  players?: Player[];
  isCoach?: boolean;
  aauNumber?: string;
  additionalGuardians?: Guardian[];
}

export interface FormData {
  fullName: string;
  email: string;
  phone: string;
  address: Address;
  relationship: string;
  isCoach: boolean;
  aauNumber: string;
}

export interface RegistrationStatus {
  parentRegistered: boolean;
  parentPaid: boolean;
  currentSeason: string;
  hasPlayers: boolean;
  hasCurrentSeasonPlayers: boolean;
  allPlayersPaid: boolean;
}

export interface AuthContextType {
  isAuthenticated: boolean;
  parent: Parent | null;
  parents: Parent[];
  user: Parent | null;
  players: Player[];
  setPlayers: React.Dispatch<React.SetStateAction<Player[]>>;
  allParents: Parent[];
  login: (email: string, password: string) => Promise<void>;
  logout: () => void;
  register: (
    email: string,
    password: string,
    fullName: string,
    phone: string,
    address: string,
    relationship: string,
    isCoach: boolean,
    aauNumber: string,
    agreeToTerms: boolean
  ) => Promise<void>;
  fetchParentData: (parentId: string) => Promise<Parent | null>;
  fetchPlayersData: (
    playerIds: string[],
    queryParams?: string
  ) => Promise<void>;
  fetchPlayerData: (playerId: string) => Promise<Player | null>;
  fetchAllPlayers: (queryParams?: string) => Promise<void>;
  fetchAllParents: (queryParams?: string) => Promise<Parent[]>;
  fetchGuardians: (playerId: string) => Promise<Guardian[]>;
  checkAuth: () => Promise<void>;
  role: string;
  isLoading: boolean;
  searchAll: (term: string) => Promise<SearchResult[]>;
  fetchParentsData: (queryParams?: string) => Promise<void>;
  fetchParentPlayers: (parentId: string) => Promise<Player[]>;
  fetchAllGuardians: (queryParams?: string) => Promise<Guardian[]>;
  currentUser: Parent | null;
  allGuardians: Guardian[];
  refreshAuthData: () => Promise<void>;
  registrationStatus: RegistrationStatus;
  setRegistrationStatus?: React.Dispatch<
    React.SetStateAction<RegistrationStatus>
  >;
}